+++ backend/db/models/spot.js
@@ -62,7 +62,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false
       },
       price: {
-        type: DataTypes.DECIMAL(7, 2),
+        type: DataTypes.FLOAT,
         allowNull: false
       }
     },
